Finally found a solution to allow streaming of HTTP Post operations. It is now possible to have POST request bodies larger than the memory size of the machine the proxy is running on.
In the end, the magic fix for this was to nest the Iteratee that
pulls from the client inside the Enumerator that pushes to the
server. This was done using a function with the following typesig:

    enumIteratee :: MonadIO m => Int64
             -> (Int64 -> Iteratee ByteString m ByteString)
             -> Enumerator ByteString (Iteratee ByteString m) c

which takes a total length to read from the Iteratee, an Iteratee
that can pull the data from the source in chunks and returns an
enumerator that works in a "Iteratee ByteString m" monad.

The other thing required was to pull the iterSocket function up from
IO to MonadIO so its typesig changes from:

    iterSocket :: T.Handle -> Socket -> Bool
                   -> E.Iteratee B.ByteString IO ()

to:
    iterSocket :: MonadIO m
           => T.Handle -> Socket -> Bool
               -> E.Iteratee B.ByteString m ()

so that it would work inside the "Iteratee ByteString IO" monad.