Network.HTTP.Proxy: Handle errors in upstream request
When adding a HTTP to HTTPS upgrade test, found that errors thrown by the
upstream http-conduit operation were not being handled and were percolating
back up to the Warp web server.

Solution was to wrap Control.Exception.handle around the http-conduit
operation with an appropriate handler to feed an error Response back to
the client.