Test.Gen: Generate a Wai.Request instead of a Proxy Request
The proxy receives a Wai.Request, converts it to the proxy's own Request
type, potentially modifies it and then converts it back to a Wai.Request.

However, the conversion from a Wai.Request to a proxy Request is lossy
so that when we do a conversion of Wai.Request to proxy Request and then
back to a Wai.Request we copy the unchanged fields from the original to
the destination.