Tests: Run test proxy on OS provided random port
Tests were failing on Travis because the random port that was chosen at the
start of the test was released and then acquired by another program.

Instead, create a localhost Socket listening on aNY_PORT and pass that Socket
to runProxySettingsSocket and get the Socket's port number to pass to the test
code.